phase_2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
better-exceptions==0.3.2
alembic==1.13.1
python-dotenv==1.0.0
better-auth[fastapi]==0.0.15
openai==1.3.5
tiktoken==0.5.2